+++ apps/client/src/app/pages/blog/blog-page-routing.module.ts
@@ -172,6 +172,15 @@ const routes: Routes = [
         './2023/11/hacktoberfest-2023-debriefing/hacktoberfest-2023-debriefing-page.component'
       ).then((c) => c.Hacktoberfest2023DebriefingPageComponent),
     title: 'Hacktoberfest 2023 Debriefing'
+  },
+  {
+    canActivate: [AuthGuard],
+    path: '2023/11/black-friday-2023',
+    loadComponent: () =>
+      import(
+        './2023/11/black-friday-2023/black-friday-2023-page.component'
+      ).then((c) => c.BlackFriday2023PageComponent),
+    title: 'Black Friday 2023'
   }
 ];